refactor: move variable helpers to common
