+++ tests/libecalc/core/consumers/conftest.py
@@ -20,9 +20,7 @@ from libecalc.domain.infrastructure.energy_components.legacy_consumer.tabulated 
     TabularConsumerFunction,
 )
 from libecalc.domain.infrastructure.path_id import PathID
-from libecalc.domain.infrastructure.energy_components.legacy_consumer.tabulated.tabular_energy_function import (
-    VariableExpression,
-)
+from libecalc.domain.infrastructure.energy_components.legacy_consumer.tabulated.common import VariableExpression
 from libecalc.domain.regularity import Regularity
 from libecalc.dto.types import ConsumerUserDefinedCategoryType
 from libecalc.expression import Expression
